feat: introduce lms blaise case model to the family
refs: BLAIS5-4351
