+++ models/update/frs_blaise_update_case_model.py
@@ -1,13 +1,13 @@
 from typing import Dict, List
 
 from enums.blaise_fields import BlaiseFields
-from models.common.blaise.blaise_case_model import BlaiseCaseModel
+from models.common.blaise.blaise_case_model import BlaiseCaseModelBase
 from models.update.totalmobile_incoming_update_request_model import (
     TotalMobileIncomingUpdateRequestModel,
 )
 
 
-class FRSBlaiseUpdateCase(BlaiseCaseModel):
+class FRSBlaiseUpdateCase(BlaiseCaseModelBase):
     def __init__(self, questionnaire_name: str, case_data: Dict[str, str]):  # type: ignore
         super().__init__(questionnaire_name, case_data)
